Add files via upload
The dataset used for the KNN , Naive Bayes and SVM model is Brain_Tumor_Dataset. It contains data regarding brain tumors, with each sample labeled as either 0 (no tumor) or 1 (tumor present). The dataset consists of 4000 samples, with an equal distribution of positive and negative cases. This data set was collected from Kaggle.  The dataset is used for both classification and regression tasks.